mt9m114: Fix compiler warning
BZ: 133029

Fixes compiler warning.

Acked-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Acked-by: Vinod Govindapillai <vinod.govindapillai@intel.com>
Reviewed-on: http://android.intel.com:8080/127391
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: cactus <cactus@intel.com>

+++ drivers/media/i2c/mt9m114.c
@@ -886,7 +886,7 @@ static int mt9m114_g_2a_status(struct v4l2_subdev *sd, s32 *val)
 static int mt9m114_g_exposure(struct v4l2_subdev *sd, s32 *value)
 {
 	struct i2c_client *client = v4l2_get_subdevdata(sd);
-	u16 coarse;
+	u32 coarse;
 	int ret;
 
 	/* the fine integration time is currently not calculated */
